+++ pkg/des.model.job.go
@@ -25,8 +25,8 @@ type DESJob struct {
 	DESJobName  string  `gorm:"not null; unique; varchar(24)" json:"des_job_name"`
 	DESJobStart int64   `gorm:"not null" json:"des_job_start"`
 	DESJobEnd   int64   `gorm:"not null" json:"des_job_end"`
-	DESJobLng   float32 `json:"des_job_lng"`
-	DESJobLat   float32 `json:"des_job_lat"`
+	DESJobLng   float64 `json:"des_job_lng"`
+	DESJobLat   float64 `json:"des_job_lat"`
 	DESJobDevID int64   `json:"des_job_dev_id"`
 
 	DESDev DESDev `gorm:"foreignKey:DESJobDevID" json:"-"`
